Squash all warnings.

+++ include/array.h
@@ -72,8 +72,8 @@ OBLCORE_IMPEXP void *        array_prev(array_t *);
 
 #define array_size(a)            ((a) -> size)
 
-#define array_set_int(a, i, v)   array_put((a), (i), (void *)((long) (v)))
-#define array_get_int(a, i)      ((long) array_get((a), (i)));
+#define array_set_int(a, i, v)   array_put((a), (i), (void *)((intptr_t) (v)))
+#define array_get_int(a, i)      ((intptr_t) array_get((a), (i)));
 #define array_push(a, d)         (array_set((a), -1, (d)))
 #define array_empty(a)           (array_size((a)) == 0)
 #define array_notempty(a)        (array_size((a)) > 0)
